kalendarz_aplikacja/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
email-validator
python-dotenv
jose
passlib
pyjwt
python-jose
python-multipart
bcrypt
npm
npx
node
sqlalchemy